New cleaner Module implementation.
