+++ equinox/_module/__init__.py
@@ -1,0 +1,14 @@
+from ._better_abstract import (
+    AbstractClassVar as AbstractClassVar,
+    AbstractVar as AbstractVar,
+)
+from ._field import (
+    field as field,
+    static_field as static_field,  # deprecated
+)
+from ._module import (
+    Module as Module,
+    module_update_wrapper as module_update_wrapper,
+    StrictConfig as StrictConfig,  # deprecated
+)
+from ._prebuilt import Partial as Partial, Static as Static
