+++ equinox/nn/_batch_norm.py
@@ -14,7 +14,7 @@ from ._sequential import StatefulLayer
 from ._stateful import State, StateIndex
 
 
-class BatchNorm(StatefulLayer, strict=True):
+class BatchNorm(StatefulLayer):
     r"""Computes a mean and standard deviation over the batch and spatial
     dimensions of an array, and uses these to normalise the whole array. Optionally
     applies a channelwise affine transformation afterwards.
