+++ equinox/nn/_shared.py
@@ -14,7 +14,7 @@ class SharedNode:
         return "SharedNode"
 
 
-class Shared(Module, strict=True):
+class Shared(Module):
     """Used to tie together multiple nodes across a PyTree.
 
     Note that Equinox modules are Py**Trees** -- so the same layer, appearing in two
