+++ equinox/nn/_weight_norm.py
@@ -21,7 +21,7 @@ def _norm_except_axis(v: Array, norm: Callable[[Array], Scalar], axis: int | Non
         return jax.vmap(norm, in_axes=axis, out_axes=axis)(v)
 
 
-class WeightNorm(Module, Generic[_Layer], strict=True):
+class WeightNorm(Module, Generic[_Layer]):
     r"""Applies weight normalisation to a given parameter.
 
     Given the 2D weight matrix
